+++ lib/monk/live/errors.rb
@@ -1,0 +1,9 @@
+module Monk
+  module Live
+    class NotFrozenError < StandardError
+    end
+
+    class NotConfiguredError < StandardError
+    end
+  end
+end